added bonus and required features

+++ schema.sql
@@ -25,8 +25,8 @@ CREATE TABLE employee
     auto_increment,
   title VARCHAR
     (50) NOT NULL,
-  salary DECIMAL
-    (10,4) NOT NULL,
+  salary INTEGER
+    (10) NOT NULL,
   department_id INT
     (20) NOT NULL,
   PRIMARY KEY
